show: clarify error message
In case of a non-existing or empty line, 'pass show --qrcode=42
passname' dies with the error message "There is no password to put
on the clipboard at line 42", even though the '--qrcode' option doesn't
involve the clipboard at all.

Update this error message to be more generic and to merely say: "There
is no password at line 42".