show: allow selecting the output line by field name
We can select a specific line from a multiline password file by
specifying its line number to the --stdout, --clip or --qrcode
options.  However, names are much more descriptive and easier to
remember than numbers, and they can be used consistently accross
multiple multiline password files.

So, let's extend the optional argument of these options to allow
selecting a line by its field name.  I.e. if the line selector
argument is a number, then select the Nth line from the password file
(so the existing behavior doesn't change), or else treat it as a field
name, and select the first line that starts with that name (followed
by ': ').