show: refactor mutually exclusive option check
'pass show's '--clip' and '--qrcode' options are mutually exclusive,
and 'pass' aborts showing usage when the user specifies both.  The
condition catching this checks whether both variables corresponding to
each of these options are set (i.e. '$qrcode -eq 1 && $clip -eq 1').
This is fine for now when there are only two mutually exclusive
options, but it won't scale when we add more such options.

Refactor this check so it scales better to arbitrary number of
mutually exclusive options: the variables corresponding to those
options have a value of either '1' or '0', so let's compute their
arithmetic sum, and error out when their sum is greater than 1.