show: add extract_one_line() helper function
A shell pipeline involving 'tail' and 'head' are responsible for
extracting a specific line from a multiline password file for
'--clip=2' and '--qrcode=3'.  With some fixes and features in upcoming
commits these pipeline stages will get more complicated, so let's
raise the level of abstraction and extract them into a dedicated
helper function.