show: add '--strip-field' option
Some services use a so called app password that, quoting Google's
support site about app passwords, "gives a less secure app or device
permission to access" the account, e.g. to access GMail with 'msmtp'
and 'mbsync'.  To handle such an app password with 'pass', it's
convenient to store the app password in the same multiline password
file as the other credentials and metadata related to that account
like this:

  abie4lohTi8geif4
  login: john.doe@gmail.com
  URL: *.google.com/*
  app-password: bae2vohnohth3Bahphei

When providing such an app password to a program that field name
prefix must be removed from the output of 'pass', of course.  While
it's fairly straightforward to do so using a pipeline and a couple of
shell commands, it's not always applicable.  Just like in the previous
commit, the problem is with restrictions imposed by AppArmor: even if
the user can configure a program to run 'pass' and the pipeline
commands to supply the app password, the program's default AppArmor
profile might only allow invoking 'pass', but prevent it from invoking
those pipeline commands necessary to strip that field name prefix.

Provide better support for those systems by adding the '--strip-field'
option to remove that field name prefix, i.e. everything up to and
including the first ': '.

  $ pass show --stdout=4 --strip-field google.com
  bae2vohnohth3Bahphei