generate: take into account character set specified in the password file
I have to deal with some services that for whatever reason allow only
a subset of non-alphanumeric characters in passwords, and, shame on
them, they don't even bother to specify the subset that they do
accept.  Figuring out the allowed special characters is always a
hassle, so I decided to record the allowed character set in the
password file itself with the prefix 'character-set:'.

To make this more useful, teach 'pass generate' to check whether there
is such a 'character-set' line in a multiline password file, and use
the character set stored within to generate the new password instead
of the default.